make max media upload size configurable
As some of us want to upload full-size photos and videos.

I’m pretty rusty with Django so sorry if this isn’t the idiomatic way to
access the setting from a view. Please let me know if not, and I’ll fix
this.

I hope to propose a follow-up change that’ll make the max image
dimensions configurable. After that, someone should be able to configure
their installation such that they’d be able to post, say, a 14 megabyte
file of a 25 megapixel photo, and have it be uploaded, saved, and
attached to a post, at the original fidelity. If so desired.