Move Nginx to separate container
By moving nginx to its own container, we make it _vastly_ easier to
insert a debugger into the environment. Once gunicorn/django is
on its own container, we can inject our debugger there without
affecting the nginx process.

This assumes that nginx and gunicorn don't have any shared
files on disk, which doesn't appear to be the case today. If such
a thing is added, a shared volume will be required.

Specifically, this unlocks trivial PyCharm debugging (and presumably
other debuggers)

Additionally:
* Create dependency tree within all containers, and specify healthcheck should complete
* Expose all container service ports for debugging
* Improve healthcheck timing on postgres container