Use cache-busting URLs for proxied files
Migrates (in a backwards-compatible way) from `/proxy/identity_image/271/` to `/proxy/identity_image/271/f5d8e72f2b/`.

The "hash tail" is generated as a truncated SHA1 of the upstream URL. This allows us to respect the `Cache-Control` header we get from upstream, as if the image changes for the identity then the hash will also change.

In other words, while the previous arrangement forced Takahē to replace the cache-headers because it didn't know when upstream changed, the new setup will immediately update the URL when the upstream does.

Overall, this gives us two benefits:
* Upstream cache settings are preserved, respecting the server's wishes
* Any upstream change will immediately propagate through all caches, including nginx and any CDNs

Nothing actually parses the hash, it exists only to bust the cache. The security of the hash doesn't matter, and the length doesn't really matter either (because it is scoped to the identity PK).

Any attempts to reach the "old" URL without a hash will continue to work, though the upstream cache-control headers will be used.

Note: If an upstream does have "bad" cache control settings, this will allow it to affect the CDN/client. While not ideal, this seems a reasonable tradeoff (and nginx will maintain its cache independently).