Templatize robots.txt
The existing static robots.txt did not allow any method of adding additional disallowed agents.

Since a substantial majority of web traffic to my test instance is a few SEO crawlers, and this is likely to be true for others as well, add a method by which well-behaving robots can be blocked.