Accept HTTP signatures where the algorithm is obfuscated at hs2019
Reference: https://datatracker.ietf.org/doc/html/draft-cavage-http-signatures-12
It is recommended to list "hs2019" instead of the actual signature algorith
in order to obscure attack vectors.

Some fediverse servers, including the library used by GoToSocial, do this,
and rejecting those signatures blocked incoming requests to the inbox.

The validation being used here can still process the signature even with
the field labeled this way, so allowing it fixes this part of the incompatibility.